pyproject.toml
[project]
name = "prediction-market-trading-bot"
version = "0.1.0"
description = "AI-powered Polymarket prediction market trading bot (paper trading MVP)"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "anthropic>=0.40.0",
    "py-clob-client-v2>=1.0.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "feedparser>=6.0.10",
    "praw>=7.7.1",
    "websockets>=12.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]
sentiment = [
    "transformers>=4.40.0",
    "torch>=2.2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
pythonpath = ["src", ".claude/skills/pm-risk/scripts", ".claude/skills/pm-predict/scripts", ".claude/skills/pm-scan/scripts", ".claude/skills/pm-research/scripts", ".claude/skills/pm-compound/scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]
